refactor:개발 코드컨벤션을 따른 코드수정
